+++ Makefile
@@ -87,16 +87,29 @@ HIVE_REVISION := df7d5103d4ddc772307f9947be4ad1f20ce03ed0
 # the whole history by making it shallow since a given date (one day before our
 # target revision).
 HIVE_SHALLOW_SINCE := 2024-09-02
+QUIET ?= false
 hive:
-	git clone --single-branch --branch master --shallow-since=$(HIVE_SHALLOW_SINCE) https://github.com/lambdaclass/hive
-	cd hive && git checkout --detach $(HIVE_REVISION) && go build .
+	if [ "$(QUIET)" = "true" ]; then \
+		git clone --quiet --single-branch --branch master --shallow-since=$(HIVE_SHALLOW_SINCE) https://github.com/lambdaclass/hive && \
+		cd hive && git checkout --quiet --detach $(HIVE_REVISION) && go build .; \
+	else \
+		git clone --single-branch --branch master --shallow-since=$(HIVE_SHALLOW_SINCE) https://github.com/lambdaclass/hive && \
+		cd hive && git checkout --detach $(HIVE_REVISION) && go build .; \
+	fi
 
 setup-hive: hive ## 🐝 Set up Hive testing framework
 	if [ "$$(cd hive && git rev-parse HEAD)" != "$(HIVE_REVISION)" ]; then \
-		cd hive && \
-		git checkout master && \
-		git fetch --shallow-since=$(HIVE_SHALLOW_SINCE) && \
-		git checkout --detach $(HIVE_REVISION) && go build . ;\
+		if [ "$(QUIET)" = "true" ]; then \
+			cd hive && \
+			git checkout --quiet master && \
+			git fetch --quiet --shallow-since=$(HIVE_SHALLOW_SINCE) && \
+			git checkout --quiet --detach $(HIVE_REVISION) && go build .;\
+		else \
+			cd hive && \
+			git checkout master && \
+			git fetch --shallow-since=$(HIVE_SHALLOW_SINCE) && \
+			git checkout --detach $(HIVE_REVISION) && go build .;\
+		fi \
 	fi
 
 TEST_PATTERN ?= /
@@ -106,7 +119,10 @@ TEST_PATTERN ?= /
 # For example, to run the rpc-compat suites for eth_chainId & eth_blockNumber you should run:
 # `make run-hive SIMULATION=ethereum/rpc-compat TEST_PATTERN="/eth_chainId|eth_blockNumber"`
 run-hive: build-image setup-hive ## 🧪 Run Hive testing suite
-	cd hive && ./hive --sim $(SIMULATION) --client ethrex --sim.limit "$(TEST_PATTERN)"
+	cd hive && ./hive --client ethrex --sim $(SIMULATION) --sim.limit "$(TEST_PATTERN)"
+
+run-hive-all: build-image setup-hive ## 🧪 Run all Hive testing suites
+	cd hive && ./hive --client ethrex --sim $(SIMULATION) --sim.parallelism 4
 
 run-hive-debug: build-image setup-hive ## 🐞 Run Hive testing suite in debug mode
 	cd hive && ./hive --sim $(SIMULATION) --client ethrex --sim.limit "$(TEST_PATTERN)" --docker.output
@@ -116,3 +132,27 @@ clean-hive-logs: ## 🧹 Clean Hive logs
 
 loc:
 	cargo run -p loc
+
+loc-stats:
+	if [ "$(QUIET)" = "true" ]; then \
+		cargo run --quiet -p loc -- --summary;\
+	else \
+		cargo run -p loc -- --summary;\
+	fi
+
+
+hive-stats:
+	make hive QUIET=true
+	make setup-hive QUIET=true
+	rm -rf hive/workspace $(FILE_NAME)_logs
+	make run-hive-all SIMULATION=ethereum/rpc-compat || exit 0
+	make run-hive-all SIMULATION=devp2p || exit 0
+	make run-hive-all SIMULATION=ethereum/engine || exit 0
+	make run-hive-all SIMULATION=ethereum/sync || exit 0
+
+stats:
+	make loc-stats QUIET=true && echo
+	cd crates/vm/levm && make download-evm-ef-tests
+	cd crates/vm/levm && make run-evm-ef-tests QUIET=true && echo
+	make hive-stats
+	cargo run --quiet --release -p hive_report
