+++ cmd/loc/Cargo.toml
@@ -7,3 +7,7 @@ edition.workspace = true
 tokei = "12.1.2"
 serde = "1.0.215"
 serde_json = "1.0.133"
+clap = { version = "4.3", features = ["derive"] }
+clap_complete = "4.5.17"
+colored = "2.1.0"
+spinoff = "0.8.0"
